feat: Moved to a more open helm-release with a targeted sops decryption. Also removed networkPolicy
The file is not so encrypted, only sensitive information.

networkPolicies are causing problems I believe they are not configured, but expect me to configure them by default. Which is ok, but at the
same time I'd expect more form the helm-chart

I need to figure out a better strategy tho... mainly for encrypting and decrypting
resources manually while testing/deving